Addression first set of comments 1] using macros to throw exception with line no. info 2] Adding variable for metadata server ip and rpc port in CMakeList.txt file
