+++ src/cis/fam_cis_direct.cpp
@@ -113,12 +113,13 @@ Fam_CIS_Direct::create_region(string name, size_t nbytes, mode_t permission,
     // Check if the name size is bigger than MAX_KEY_LEN supported
     if (name.size() > metadataManager->metadata_maxkeylen()) {
         message << "Name too long";
-        throw CIS_Exception(REGION_NAME_TOO_LONG, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NAME_TOO_LONG,
+                        message.str().c_str());
     }
 
     if (metadataManager->metadata_find_region(name, region)) {
         message << "Region already exist";
-        throw CIS_Exception(REGION_EXIST, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_EXIST, message.str().c_str());
     }
     allocator->create_region(name, regionId, nbytes, permission, uid, gid);
 
@@ -146,7 +147,7 @@ void Fam_CIS_Direct::destroy_region(uint64_t regionId, uint64_t memoryServerId,
     Fam_Region_Metadata region;
     if (!(metadataManager->metadata_find_region(regionId, region))) {
         message << "Region does not exist";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     // Check if calling PE user is owner. If not, check with
@@ -157,8 +158,8 @@ void Fam_CIS_Direct::destroy_region(uint64_t regionId, uint64_t memoryServerId,
             &region, META_REGION_ITEM_WRITE, uid, gid);
         if (!isPermitted) {
             message << "Destroying region is not permitted";
-            throw CIS_Exception(DESTROY_REGION_NOT_PERMITTED,
-                                message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, DESTROY_REGION_NOT_PERMITTED,
+                            message.str().c_str());
         }
     }
 
@@ -186,14 +187,15 @@ void Fam_CIS_Direct::resize_region(uint64_t regionId, size_t nbytes,
     Fam_Region_Metadata region;
     if (!(metadataManager->metadata_find_region(regionId, region))) {
         message << "Region does not exist";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     bool isPermitted = metadataManager->metadata_check_permissions(
         &region, META_REGION_ITEM_WRITE, uid, gid);
     if (!isPermitted) {
         message << "Region resize not permitted";
-        throw CIS_Exception(REGION_RESIZE_NOT_PERMITTED, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_RESIZE_NOT_PERMITTED,
+                        message.str().c_str());
     }
 
     allocator->resize_region(regionId, uid, gid, nbytes);
@@ -220,14 +222,15 @@ Fam_Region_Item_Info Fam_CIS_Direct::allocate(string name, size_t nbytes,
     // Check if the name size is bigger than MAX_KEY_LEN supported
     if (name.size() > metadataManager->metadata_maxkeylen()) {
         message << "Name too long";
-        throw CIS_Exception(DATAITEM_NAME_TOO_LONG, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NAME_TOO_LONG,
+                        message.str().c_str());
     }
 
     // Check with metadata service if the region exist, if not return error
     Fam_Region_Metadata region;
     if (!metadataManager->metadata_find_region(regionId, region)) {
         message << "Region does not exist";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     // Check if calling PE user is owner. If not, check with
@@ -238,8 +241,8 @@ Fam_Region_Item_Info Fam_CIS_Direct::allocate(string name, size_t nbytes,
             &region, META_REGION_ITEM_WRITE, uid, gid);
         if (!isPermitted) {
             message << "Allocation of dataitem is not permitted";
-            throw CIS_Exception(DATAITEM_ALLOC_NOT_PERMITTED,
-                                message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, DATAITEM_ALLOC_NOT_PERMITTED,
+                            message.str().c_str());
         }
     }
 
@@ -250,7 +253,8 @@ Fam_Region_Item_Info Fam_CIS_Direct::allocate(string name, size_t nbytes,
     if (name != "") {
         if (metadataManager->metadata_find_dataitem(name, regionId, dataitem)) {
             message << "Dataitem with the name provided already exist";
-            throw CIS_Exception(DATAITEM_EXIST, message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, DATAITEM_EXIST,
+                            message.str().c_str());
         }
     }
 
@@ -307,7 +311,8 @@ void Fam_CIS_Direct::deallocate(uint64_t regionId, uint64_t offset,
     if (!metadataManager->metadata_find_dataitem(dataitemId, regionId,
                                                  dataitem)) {
         message << "Dataitem does not exist";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     // Check if calling PE user is owner. If not, check with
@@ -318,8 +323,8 @@ void Fam_CIS_Direct::deallocate(uint64_t regionId, uint64_t offset,
             &dataitem, META_REGION_ITEM_WRITE, uid, gid);
         if (!isPermitted) {
             message << "Deallocation of dataitem is not permitted";
-            throw CIS_Exception(DATAITEM_DEALLOC_NOT_PERMITTED,
-                                message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, DATAITEM_DEALLOC_NOT_PERMITTED,
+                            message.str().c_str());
         }
     }
 
@@ -346,15 +351,15 @@ void Fam_CIS_Direct::change_region_permission(uint64_t regionId,
     Fam_Region_Metadata region;
     if (!metadataManager->metadata_find_region(regionId, region)) {
         message << "Region does not exist";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     // Check with metadata service if the calling PE has the permission
     // to modify permissions of the region, if not return error
     if (uid != region.uid) {
         message << "Region permission modification not permitted";
-        throw CIS_Exception(REGION_PERM_MODIFY_NOT_PERMITTED,
-                            message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_PERM_MODIFY_NOT_PERMITTED,
+                        message.str().c_str());
     }
 
     // Update the permission of region with metadata service
@@ -381,15 +386,16 @@ void Fam_CIS_Direct::change_dataitem_permission(uint64_t regionId,
     if (!metadataManager->metadata_find_dataitem(dataitemId, regionId,
                                                  dataitem)) {
         message << "Dataitem does not exist";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     // Check with metadata service if the calling PE has the permission
     // to modify permissions of the region, if not return error
     if (uid != dataitem.uid) {
         message << "Dataitem permission modification not permitted";
-        throw CIS_Exception(ITEM_PERM_MODIFY_NOT_PERMITTED,
-                            message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, ITEM_PERM_MODIFY_NOT_PERMITTED,
+                        message.str().c_str());
     }
 
     // Update the permission of region with metadata service
@@ -445,13 +451,14 @@ Fam_Region_Item_Info Fam_CIS_Direct::lookup_region(string name,
     message << "Error While locating region : ";
     if (!metadataManager->metadata_find_region(name, region)) {
         message << "could not find the region";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     if (uid != region.uid) {
         if (!check_region_permission(region, 0, uid, gid)) {
             message << "Not permitted to access the region";
-            throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION,
+                            message.str().c_str());
         }
     }
     info.regionId = region.regionId;
@@ -475,13 +482,15 @@ Fam_Region_Item_Info Fam_CIS_Direct::lookup(string itemName, string regionName,
     if (!metadataManager->metadata_find_dataitem(itemName, regionName,
                                                  dataitem)) {
         message << "could not find the dataitem ";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     if (uid != dataitem.uid) {
         if (!check_dataitem_permission(dataitem, 0, uid, gid)) {
             message << "Not permitted to access the dataitem";
-            throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION,
+                            message.str().c_str());
         }
     }
 
@@ -505,13 +514,14 @@ Fam_Region_Item_Info Fam_CIS_Direct::check_permission_get_region_info(
     message << "Error While locating region : ";
     if (!metadataManager->metadata_find_region(regionId, region)) {
         message << "could not find the region";
-        throw CIS_Exception(REGION_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, REGION_NOT_FOUND, message.str().c_str());
     }
 
     if (uid != region.uid) {
         if (!check_region_permission(region, 0, uid, gid)) {
             message << "Not permitted to access the region";
-            throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION,
+                            message.str().c_str());
         }
     }
     info.size = region.size;
@@ -537,7 +547,8 @@ Fam_Region_Item_Info Fam_CIS_Direct::check_permission_get_item_info(
     if (!metadataManager->metadata_find_dataitem(dataitemId, regionId,
                                                  dataitem)) {
         message << "could not find the dataitem";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     // This key is only applicable for shared memory model
@@ -549,7 +560,7 @@ Fam_Region_Item_Info Fam_CIS_Direct::check_permission_get_item_info(
         key = FAM_KEY_INVALID;
     } else {
         message << "Dataitem access is not permitted ";
-        throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION, message.str().c_str());
     }
 
     info.regionId = dataitem.regionId;
@@ -579,13 +590,15 @@ Fam_Region_Item_Info Fam_CIS_Direct::get_stat_info(uint64_t regionId,
     if (!metadataManager->metadata_find_dataitem(dataitemId, regionId,
                                                  dataitem)) {
         message << "could not find the dataitem";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     if (uid != dataitem.uid) {
         if (!check_dataitem_permission(dataitem, 0, uid, gid)) {
             message << "Not permitted to access the dataitem";
-            throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+            THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION,
+                            message.str().c_str());
         }
     }
 
@@ -612,14 +625,15 @@ void *Fam_CIS_Direct::fam_map(uint64_t regionId, uint64_t offset,
     if (!metadataManager->metadata_find_dataitem(dataitemId, regionId,
                                                  dataitem)) {
         message << "could not find the dataitem";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
     if (check_dataitem_permission(dataitem, 1, uid, gid) |
         check_dataitem_permission(dataitem, 0, uid, gid)) {
         localPointer = get_local_pointer(regionId, offset);
     } else {
-        throw CIS_Exception(NO_PERMISSION,
-                            "Not permitted to use this dataitem");
+        THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION,
+                        "Not permitted to use this dataitem");
     }
     CIS_DIRECT_PROFILE_END_OPS(cis_fam_map);
 
@@ -649,45 +663,48 @@ void *Fam_CIS_Direct::copy(uint64_t srcRegionId, uint64_t srcOffset,
     if (!metadataManager->metadata_find_dataitem(srcDataitemId, srcRegionId,
                                                  srcDataitem)) {
         message << "could not find the source dataitem";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     if (!(metadataManager->metadata_check_permissions(
             &srcDataitem, META_REGION_ITEM_READ, uid, gid))) {
         message << "Read operation is not permitted on source dataitem";
-        throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION, message.str().c_str());
     }
 
     if (!metadataManager->metadata_find_dataitem(destDataitemId, destRegionId,
                                                  destDataitem)) {
         message << "could not find the destination dataitem";
-        throw CIS_Exception(DATAITEM_NOT_FOUND, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, DATAITEM_NOT_FOUND,
+                        message.str().c_str());
     }
 
     if (!(metadataManager->metadata_check_permissions(
             &destDataitem, META_REGION_ITEM_WRITE, uid, gid))) {
         message << "Write operation is not permitted on destination dataitem";
-        throw CIS_Exception(NO_PERMISSION, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, NO_PERMISSION, message.str().c_str());
     }
 
     if ((srcCopyStart + nbytes) < srcDataitem.size)
         srcStart = get_local_pointer(srcRegionId, srcOffset + srcCopyStart);
     else {
         message << "Source offset or size is beyond dataitem boundary";
-        throw CIS_Exception(OUT_OF_RANGE, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, OUT_OF_RANGE, message.str().c_str());
     }
 
     if ((destCopyStart + nbytes) < destDataitem.size)
         destStart = get_local_pointer(destRegionId, destOffset + destCopyStart);
     else {
         message << "Destination offset or size is beyond dataitem boundary";
-        throw CIS_Exception(OUT_OF_RANGE, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, OUT_OF_RANGE, message.str().c_str());
     }
 
     if ((srcStart == NULL) || (destStart == NULL)) {
         message
             << "Failed to get local pointer to source or destination dataitem";
-        throw CIS_Exception(NULL_POINTER_ACCESS, message.str().c_str());
+        THROW_ERRNO_MSG(CIS_Exception, NULL_POINTER_ACCESS,
+                        message.str().c_str());
     } else {
         allocator->copy(destStart, srcStart, nbytes);
     }
