+++ src/common/fam_internal_exception.cpp
@@ -60,4 +60,67 @@ Metadata_Service_Exception::Metadata_Service_Exception(enum Fam_Error serverErr,
     famErrMsg = msg;
     famErr = serverErr;
 }
+
+Fam_Permission_Exception::Fam_Permission_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_NOPERM;
+}
+
+Fam_Permission_Exception::Fam_Permission_Exception()
+    : Fam_Permission_Exception("No Permission") {}
+
+Fam_InvalidOption_Exception::Fam_InvalidOption_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_INVALID;
+}
+
+Fam_InvalidOption_Exception::Fam_InvalidOption_Exception()
+    : Fam_InvalidOption_Exception("Invalid Option") {}
+
+Fam_Timeout_Exception::Fam_Timeout_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_TIMEOUT;
+}
+
+Fam_Timeout_Exception::Fam_Timeout_Exception()
+    : Fam_Timeout_Exception("Timeout failure") {}
+
+Fam_Datapath_Exception::Fam_Datapath_Exception(enum Fam_Error fErr,
+                                               const char *msg) {
+    famErrMsg = msg;
+    famErr = fErr;
+}
+
+Fam_Datapath_Exception::Fam_Datapath_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_LIBFABRIC;
+}
+
+Fam_Datapath_Exception::Fam_Datapath_Exception()
+    : Fam_Datapath_Exception("Datapath failure") {}
+
+Fam_Allocator_Exception::Fam_Allocator_Exception(enum Fam_Error fErr,
+                                                 const char *msg) {
+    famErrMsg = msg;
+    famErr = fErr;
+}
+
+Fam_Allocator_Exception::Fam_Allocator_Exception()
+    : Fam_Allocator_Exception(FAM_ERR_ALLOCATOR, "Allocator failure") {}
+
+Fam_Pmi_Exception::Fam_Pmi_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_PMI;
+}
+
+Fam_Pmi_Exception::Fam_Pmi_Exception() : Fam_Pmi_Exception("PMI failure") {}
+
+Fam_Unimplemented_Exception::Fam_Unimplemented_Exception(const char *msg) {
+    famErrMsg = msg;
+    famErr = FAM_ERR_UNIMPL;
+}
+
+Fam_Unimplemented_Exception::Fam_Unimplemented_Exception()
+    : Fam_Unimplemented_Exception("Function Not implemented") {}
+
 } // namespace openfam
