+++ src/metadata/fam_metadata_manager_direct.h
@@ -1,7 +1,7 @@
 /*
  *   fam_metadata_manager_direct.h
  *   Copyright (c) 2019-2020 Hewlett Packard Enterprise Development, LP. All
- * rights reserved. Redistribution and use in source and binary forms, with or
+ *   rights reserved. Redistribution and use in source and binary forms, with or
  *   without modification, are permitted provided that the following conditions
  *   are met:
  *   1. Redistributions of source code must retain the above copyright notice,
