+++ test/common/fam_test_config.h.in
@@ -42,7 +42,9 @@ using namespace openfam;
 // Fam_Options
 #cmakedefine TEST_DEFAULT_REGION "${TEST_DEFAULT_REGION}"
 #cmakedefine TEST_MEMORY_SERVER "${TEST_MEMORY_SERVER}"
+#cmakedefine TEST_METADATA_SERVER "${TEST_METADATA_SERVER}"
 #cmakedefine TEST_GRPC_PORT "${TEST_GRPC_PORT}"
+#cmakedefine TEST_METADATA_RPC_PORT "${TEST_METADATA_RPC_PORT}"
 #cmakedefine TEST_LIBFABRIC_PORT "${TEST_LIBFABRIC_PORT}"
 #cmakedefine TEST_LIBFABRIC_PROVIDER "${TEST_LIBFABRIC_PROVIDER}"
 #cmakedefine TEST_FAM_THREAD_MODEL "${TEST_FAM_THREAD_MODEL}"
