+++ test/microbench/fam-api-mb/fam_microbenchmark_datapath.cpp
@@ -299,7 +299,7 @@ int main(int argc, char **argv) {
     }
 
 #if !defined(SHM) && defined(MEMSERVER_PROFILE)
-    const char *memoryServers = strdup(TEST_MEMSERVER_IP);
+    const char *memoryServers = strdup(TEST_MEMORY_SERVER);
     std::string delimiter1 = ",";
     std::string delimiter2 = ":";
     CISServerMap memoryServerList =
